requirements.txt
flask
flask-cors
opencv-contrib-python
numpy
Pillow